% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.scheduleR_config}
\alias{print.scheduleR_config}
\title{Print a scheduleR_config or save it to a file as JSON}
\usage{
\method{print}{scheduleR_config}(x, toJSON = FALSE, file = tempfile(), ...)
}
\arguments{
\item{x}{an object of class scheduleR_config}

\item{toJSON}{logical, indicating to save the scheduleR_config as JSON in a file}

\item{file}{path to the file where to save the config a JSON}

\item{...}{other arguments, not used yet}
}
\value{
invisible()
}
\description{
Print a scheduleR_config or save it to a file as JSON.
}
\examples{
x <- scheduleR_config()
x
f <- tempfile()
print(x, toJSON=TRUE, file = f)

jsonlite::fromJSON(f, simplifyVector = FALSE)
}
\seealso{
\code{\link{scheduleR_config}}
}

