% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{RscheduleR-package}
\alias{RscheduleR}
\alias{RscheduleR-package}
\title{Interface to scheduleR for scheduling R scripts, Markdown reports and Shiny apps.}
\description{
Interface to scheduleR for scheduling R scripts, Markdown reports and Shiny apps. Sets up the default user, and creates an RStudio addin.
}

