% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{class}
\name{scheduleR-class}
\alias{scheduleR}
\alias{scheduleR-class}
\title{Connect to a scheduleR instance.}
\description{
Connect to a scheduleR instance.
}
\section{Fields}{

\describe{
\item{\code{host}}{character string with the address of the mongodb server where scheduleR is running. E.g. 127.0.0.1:27017. Change the IP address accordingly.}

\item{\code{db}}{character string with the name of mongodb database where scheduleR stores its data. Defaults to 'scheduleR'.}

\item{\code{username}}{character string with the username of the mongodb database. Use only if you have changed the Docker settings.}

\item{\code{password}}{character string with the password of the mongodb database. Use only if you have changed the Docker settings.}
}}
\section{Methods}{

\describe{
\item{\code{initialize(host = "127.0.0.1", port = 27017L, db = "scheduleR",
  username = "", password = "", rapacheport = 3080L)}}{Set up the connection to the MongoDB scheduleR database}

\item{\code{task_create(file, name = "task_identifier",
  description = "R script description", cron = "* * * * *", user,
  arguments = "", mailOnSuccess = "", mailOnError = "",
  ignoreLock = FALSE, enabled = TRUE, debug = FALSE)}}{Create a specific task which will be run using a specific cron schedule. 
Give the path the the R script, the name of the task, a description and when it has to be run. 
The task will be run by the user you provided. 
Optionally set emails which to send log messages to in case of success or failure.}

\item{\code{task_delete(name)}}{Delete a task with a certain name}

\item{\code{tasks()}}{Get a data frame with all scheduled tasks}

\item{\code{users()}}{Get a data frame with all users of scheduleR}

\item{\code{users_add_guest(email = "info@bnosac.be", password = "L@unch321")}}{Add a guest user - guest user has username 'guest', password 'L@unch321' and the email address you provide}

\item{\code{users_delete_guest()}}{Delete the guest user}
}}
\examples{
\dontrun{
x <- new("scheduleR", host = "localhost", port = 27017L, db = "scheduleR")
x
x$tasks()
x$users()
x$users_delete_guest()
x$users_add_guest()
x$users()
myscript <- system.file("extdata", "helloworld.R", package = "RscheduleR")
x$task_create(file = myscript, user = "guest", name = "mysimulation", 
   cron = "* * * * *", 
   mailOnSuccess = "computersaysyes@bnosac.be", mailOnError = "computersaysno@bnosac.be")
x$tasks()
x$task_delete(name = "mysimulation")
}
}

