% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{class}
\name{scheduleR}
\alias{scheduleR}
\title{Connect to a scheduleR instance.}
\arguments{
\item{host}{character string with the address of the mongodb server where scheduleR is running. E.g. 127.0.0.1:27017. Change the IP address accordingly.}

\item{db}{character string with the name of mongodb database where scheduleR stores its data. Defaults to 'scheduleR'.}

\item{username}{character string with the username of the mongodb database. Use only if you have changed the Docker settings.}

\item{password}{character string with the password of the mongodb database. Use only if you have changed the Docker settings.}
}
\description{
Connect to a scheduleR instance.
}
\section{Methods}{

\describe{
  \item{\code{tasks()}}{Get a list of tasks (R scripts) which are scheduled at the scheduleR instance.}
  \item{\code{task_delete(name)}}{Remove a task which was scheduled at the scheduleR instance. The name is the name of the task as returned by tasks()}
  \item{\code{task_create(file, ...)}}{Schedule an R script at a specific timepoint at the scheduleR instance.}
}
}
\examples{
\dontrun{
x <- scheduleR(host = "localhost", port = 27017L, db = "scheduleR")
x
x$tasks()
x$users()
x$users_delete_guest()
x$users_add_guest()
x$users()
myscript <- system.file("extdata", "helloworld.R", package = "RscheduleR")
x$task_create(file = myscript, user = "guest", name = "mysimulation", 
   cron = "* * * * *", 
   mailOnSuccess = "computersaysyes@bnosac.be", mailOnError = "computersaysno@bnosac.be")
x$tasks()
x$task_delete(name = "mysimulation")
}
}

