% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{scheduleR_config}
\alias{scheduleR_config}
\title{Creates a user.config.json file of a scheduleR instance.}
\usage{
scheduleR_config(uploadDir = "/home/scheduler/scripts",
  errorNotificationMailAddresses, mailer,
  mailer.from = "scheduleR <computersaysno@bnosac.be>",
  mailer.auth.user = "please.changeme@google.be",
  mailer.auth.pass = "fillinyourgooglepwd", db, db.host = "localhost",
  db.suffix = "scheduleR", port = 3000L)
}
\arguments{
\item{uploadDir}{scheduleR path to upload your scripts to at the scheduleR server. Defaults to /home/scheduler/scripts}

\item{errorNotificationMailAddresses}{character vector of email addresses that will be added to the recipients of error notification mails}

\item{mailer}{a full mailer list object if not specified through mailer.from, mailer.auth.user or mailer.auth.pass}

\item{mailer.from}{the google mail address which will be shown if a mail is sent out from scheduleR in case of success/failure}

\item{mailer.auth.user}{a google mail address used to send out the notification/report mails from. scheduleR uses nodemailer: https://github.com/nodemailer/nodemailer}

\item{mailer.auth.pass}{the password of the mailer.auth.user email}

\item{db}{a full db list object if not specified through db.url and db.suffix}

\item{db.host}{character string with the name of the url where the mongo database is running (the backbone of scheduler). Defaults to localhost. Change to the IP address where you host mongo.}

\item{db.suffix}{character string with the name of the mongo document database where logs, tasks, users, ... of the scheduleR instance are stored.}

\item{port}{port where the scheduleR application runs under. Defaults to 3000.}
}
\value{
a list with scheduleR configuration settings which can be used by the scheduleR application
}
\description{
Creates a user.config.json file of a scheduleR instance.
This file contains information about the mongo database and how to mail results of scheduleR.
}
\examples{
x <- scheduleR_config()
str(x)
x <- scheduleR_config(db.host = "19.19.19.19", 
   mailer.from = "scheduleR <computersaysno@bnosac.be>",
   mailer.auth.user = "computersaysno@bnosac.be", 
   mailer.auth.pass = "abc123")
str(x)
jsonlite::toJSON(x)
}

